+++ src/pages/performance/server/connect.js
@@ -6,7 +6,7 @@ const connect = () => {
     mongoose.set("debug", true)
   }
   mongoose.connect(connectionUrl,{
-    dbName : "performance"
+    dbName : "playtopia"
   })
   .then(()=>{
     console.log("Connected to MongoDB")
